Use a cli tool to simplify command-line parsing
This allows a simpler code, and also to produce a documentation
with `--help` for each commands.

Note: this is a breaking change, as it is not possible
to reuse previous command flags.

Note: one benefit is that it is now possible to specify
the winrm password via an environment variable.