Fix #101 - winrm-cli blocking reading stdin
Due to a change in winrm, winrm is now waiting properly on stdin
that something becomes readable.
Unfortunately when run in winrm-cli if stdin is a terminal there's
nothing to read and the command blocks.
winrm now detects stdin is a terminal, and in this case doesn't tell
winrm to expect inputs.